# api/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Sea Music API",
    "version": "0.1.0",
    "description": "Versioned HTTP contract for the UGC video platform backend."
  },
  "servers": [
    {
      "url": "http://127.0.0.1:8080",
      "description": "Local API base URL"
    }
  ],
...
  "paths": {
...
    "/api/v1/users": {
      "post": {
        "operationId": "registerUser",
        "summary": "Register an account",
        "tags": ["Identity"],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RegisterRequest"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Account registered.",
            "headers": {
              "X-Request-ID": {
                "$ref": "#/components/headers/RequestID"
              }
            },
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UserResponse"
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/BadRequest"
          },
          "409": {
            "$ref": "#/components/responses/Conflict"
          },
          "429": {
            "$ref": "#/components/responses/TooManyRequests"
          }
        }
      }
    },
    "/api/v1/sessions": {
      "post": {
        "operationId": "login",
        "summary": "Create a rotating session",
        "tags": ["Identity"],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/LoginRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Credentials accepted and session created.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TokenPair"
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/BadRequest"
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "429": {
            "$ref": "#/components/responses/TooManyRequests"
          }
        }
      }
    },
    "/api/v1/sessions/refresh": {
      "post": {
        "operationId": "refreshSession",
        "summary": "Rotate a refresh token",
        "tags": ["Identity"],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RefreshRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Refresh token rotated.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TokenPair"
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/BadRequest"
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "429": {
            "$ref": "#/components/responses/TooManyRequests"
          }
        }
      }
    },
    "/api/v1/me": {
      "get": {
        "operationId": "getCurrentUser",
        "summary": "Get the authenticated user",
        "tags": ["Identity"],
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "Authenticated account.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UserResponse"
                }
              }
            }
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "429": {
            "$ref": "#/components/responses/TooManyRequests"
          }
        }
      }
    },
...
  },
  "components": {
...
    "headers": {
      "RequestID": {
        "description": "Request correlation identifier. A valid client value is preserved; otherwise the server generates one.",
        "schema": {
          "type": "string",
          "minLength": 1,
          "maxLength": 128
        }
      }
    },
    "responses": {
      "BadRequest": {
        "description": "The request is syntactically valid HTTP but fails API validation.",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/ErrorResponse"
            }
          }
        }
      },
      "Unauthorized": {
        "description": "Authentication credentials are missing or invalid.",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/ErrorResponse"
            }
          }
        }
      },
...
      "Conflict": {
        "description": "The request conflicts with current resource state or uniqueness constraints.",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/ErrorResponse"
            }
          }
        }
      },
      "TooManyRequests": {
        "description": "The caller exceeded a rate limit.",
        "headers": {
          "Retry-After": {
            "description": "Seconds until the caller can retry.",
            "schema": {
              "type": "integer",
              "minimum": 1
            }
          }
        },
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/ErrorResponse"
            }
          }
        }
      },
...
    },
    "schemas": {
...
      "APIError": {
        "type": "object",
        "additionalProperties": false,
        "required": ["code", "message"],
        "properties": {
          "code": {
            "type": "string",
            "pattern": "^[a-z][a-z0-9_]*$",
            "examples": ["not_found"]
          },
          "message": {
            "type": "string",
            "minLength": 1
          },
          "details": {
            "type": "object",
            "additionalProperties": true
          }
        }
      },
      "ErrorResponse": {
        "type": "object",
        "additionalProperties": false,
        "required": ["error", "request_id"],
        "properties": {
          "error": {
            "$ref": "#/components/schemas/APIError"
          },
          "request_id": {
            "type": "string",
            "minLength": 1,
            "maxLength": 128
          }
        }
      },
...
      "RegisterRequest": {
        "type": "object",
        "additionalProperties": false,
        "required": ["username", "email", "password"],
        "properties": {
          "username": {
            "type": "string",
            "pattern": "^[A-Za-z0-9_]{3,32}$"
          },
          "email": {
            "type": "string",
            "format": "email",
            "maxLength": 254
          },
          "password": {
            "type": "string",
            "minLength": 12,
            "maxLength": 128,
            "writeOnly": true
          }
        }
      },
      "LoginRequest": {
        "type": "object",
        "additionalProperties": false,
        "required": ["identity", "password"],
        "properties": {
          "identity": {
            "type": "string",
            "minLength": 1,
            "maxLength": 254
          },
          "password": {
            "type": "string",
            "maxLength": 128,
            "writeOnly": true
          }
        }
      },
      "RefreshRequest": {
        "type": "object",
        "additionalProperties": false,
        "required": ["refresh_token"],
        "properties": {
          "refresh_token": {
            "type": "string",
            "minLength": 32,
            "maxLength": 512,
            "writeOnly": true
          }
        }
      },
      "TokenPair": {
        "type": "object",
        "additionalProperties": false,
        "required": ["access_token", "refresh_token", "token_type", "expires_in"],
        "properties": {
          "access_token": {
            "type": "string",
            "writeOnly": true
          },
          "refresh_token": {
            "type": "string",
            "writeOnly": true
          },
          "token_type": {
            "type": "string",
            "const": "Bearer"
          },
          "expires_in": {
            "type": "integer",
            "minimum": 1
          }
        }
      },
      "User": {
        "type": "object",
        "additionalProperties": false,
        "required": ["id", "username", "email", "role", "created_at"],
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "username": {
            "type": "string"
          },
          "email": {
            "type": "string",
            "format": "email"
          },
          "role": {
            "type": "string",
            "enum": ["member", "moderator", "admin"]
          },
          "created_at": {
            "type": "string",
            "format": "date-time"
          }
        }
      },
      "UserResponse": {
        "type": "object",
        "additionalProperties": false,
        "required": ["user"],
        "properties": {
          "user": {
            "$ref": "#/components/schemas/User"
          }
        }
      },
...
    },
    "securitySchemes": {
      "BearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    }
  }
}